requirements.txt
twilio
flask
dotenv
firebase-admin
google-genai